ciao-deploy: Abstract non-anchor cert creation into utils
This will allow the reuse of non-anchor certificate generation for
controller and launcher as well as CNCI agent.
